Changes to clean up Xfractint compiler warnings.
git-svn-id: https://svn.fractint.net/trunk/fractint@1553 cff58d23-3d7a-4071-9076-c60e8c9d8568

+++ headers/fractint.h
@@ -29,7 +29,7 @@ typedef unsigned USEGTYPE;
 #else
 typedef char * SEGTYPE;
 typedef char * USEGTYPE;
-#   define MK_FP(seg,off) (VOIDFARPTR )(seg+off)
+#   define MK_FP(seg,off) (VOIDFARPTR )((long)seg + (long)off)
 #include <sys/types.h> /* need size_t */
 #endif
